Feat: Challenge 기능 구현을 위한 Controller, Repository, Service 수정

+++ src/main/java/com/kuit/moamoa/dto/response/challenge/PublicChallengeResponse.java
@@ -13,7 +13,7 @@ public class PublicChallengeResponse {
     private Long challengeId;
     private String title;
     private String content;
-    private Integer duration;
+    private Long duration;
     private Integer battleCoin;
     private Boolean publicChallenge;
     private LocalDateTime startDate;
